Retrieve and summarize result data
In addition to the scan data, the cxInsight_9_0.ps1 script now also
retrieves and summarizes result data. This data is written to a
separate file so the script now, as its output, generates a zip
archive containing both the scan and the result data.

The getScanOdata function has been changed to accept the output
filename as its parameter so that this filename need only be specified
in the one place (the new getResultOData function is the same).