Add result triage data to cxInsight_8_9.ps1
This actually involved copying the 9.0 version of the script and then
making the changes necessary for it to work with CxSAST 8.9. The main
difference is the authentication mechanism: 8.9 uses HTTP basic
authorization via the -Credential option to the Invoke-RestMethod
cmdlet.

The odata functionality has been refactored to simplify the code (mainly
to add support for the -AllowUnencryptedAuthentication option which is
useful in test environments but hopefully less so in production
environments).