SAST-Create-Report: allow report type to be given on command line
A new command line option has been added, report_type, which allows
the user to specify the report type (CSV, PDF, RTF or XML) on the
command line (as opposed to editing the CxReportTemplate.psd1 file).

Also, the filename suffix of the generated reports now matches the
selected report type (previously it was always ".pdf").