SAST scans report - single large CSV output
The report produces a CSV with the standard SAST CSV report format, with the additon of the team path and project name as columns.  The latest scan for each project in the teams provided will be included in this single CSV, each scan results list getting appended to the file.  The teams are supplied by ReportTeams.txt file, one line per team path.  The batch file included shows the example command to run the PowerShell.  To run the batch file supply the SAST URL and direct log output to a file.