Support scan tags that differ only by case
In PowerShell, object attributes are not case sensitive. This means
that if a scan has tags that differ only by case (e.g., "TEST" and
"Test"), when the Invoke-RestMethod cmdlet tries to convert the JSON
API response to a PowerShell object the conversion fails and the
cmdlet returns the API response as a string instead. With this change,
we use the ConvertFrom-Json cmdlet to convert that string to a
HashTable which does allow keys that differ only in case. Furthermore,
as PowerShell allows HashTables to be accessed as though they were
objects, the change is relatively unintrusive.

The script now exits if the PowerShell version is less than
7. Technically, it should accept PowerShell 6 but I strongly doubt
that anyone actually uses PowerShell 6. The version installed by
default on modern Windows systems is 5.1, and I can't imagine anyone
installing version 6 given that we are now up to version 7.4.0.