EHC: Fix support for critical severity (#78)
* Only retrieve critical severity count if on SAST 9.7

* Only retrieve critical severity count if on SAST 9.7

If a non-existent property (e.g., "Critical" for CxSAST versions prior
to 9.7) is included in the OData scan query, the query
options ($expand, $filter and $select) are ignored.

This change uses the /cxrestapi/system/version REST API endpoint to
retrieve the CxSAST version which is then used to conditionally
include the "Critical" property in the request for the scan data.