Use the System.URI class for all URIs
A side-effect of switching to using the System.URI class for the
$cx_sast_server variable is that the string representation has a
trailing slash (i.e., the default path of the URI is the root of the
server). E.g.:

http://example.com/

Previously, we just concatenated this with the desired path leading to
a repeated forward slash. E.g.:

http://example.com//cxrestapi/auth/identity/connect/token

In general, this seems not to have caused any problems, but for one
user it has resulted in the call to retrieve the access token
generating a 404 response (Not Found).

While the simple solution would have been to make sure that none of
the concatenation statements introduced a spurious forward slash, the
proper solution is to use the System.URI class everywhere which is
what this change does.